Reformat

+++ src/main/java/net/drgmes/dwm/items/screwdriver/screens/ScrewdriverInterfaceMainScreen.java
@@ -19,7 +19,7 @@ public class ScrewdriverInterfaceMainScreen extends BaseScrewdriverInterfaceScre
     @Override
     protected void init() {
         int buttonWidth = (int) (this.getBackgroundSize().x - BACKGROUND_BORDERS * 2 - PADDING * 2);
-        int buttonOffset = (int) (BACKGROUND_BORDERS + PADDING);
+        int buttonOffset = BACKGROUND_BORDERS + PADDING;
 
         Vec2 modeButtonPos = this.getRenderPos(BACKGROUND_BORDERS + PADDING, buttonOffset);
         this.modeButton = new Button((int) modeButtonPos.x, (int) modeButtonPos.y, buttonWidth, BUTTON_HEIGHT, this.getModeButtonTitle(), (btn) -> {
